policy-engine/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.20
langchain-google-genai==1.0.3
langchain-community==0.0.38
chromadb==0.5.0
pypdf==4.2.0
google-generativeai==0.5.4
sentence-transformers==2.7.0
python-dotenv==1.0.1
requests==2.31.0
pytest==8.1.0
